Don't make our callers have fatal warnings
Role::Tiny advertises itself as a "slice of Moose", but Moose roles do
not import fatal warnings on you:

    $ perl -wle 'use warnings FATAL => "all"; use warnings;
    print "foo" + undef; print "bar"'
    Use of uninitialized value in addition (+) at -e line 1.
    $ perl -wle 'use warnings FATAL => "all"; no warnings;
    use warnings; print "foo" + undef; print "bar"'
    Use of uninitialized value in addition (+) at -e line 1.
    Argument "foo" isn't numeric in addition (+) at -e line 1.
    0
    bar

This isn't documented in Role::Tiny, is very annoying because code not
expecting fatal warnings being converted/written with the same
assumptions as Moose roles will now die on something that should be a
harmless warning.

So treat this as a bug and just make Role::Tiny use the default
Moose::Role behavior.